Connection reset when trying to query HS2 from Docker host
Connections with HS2 are established correctly if we are inside the
Docker container but not when we try to connect from the Docker
host.

Change hive.server2.thrift.bind.host property from localhost
to 0.0.0.0 to accept connections in all interfaces and not only
in the loopback interface, which works only if we are inside
the container.

References:
https://stackoverflow.com/questions/54940990/docker-cant-connect-to-container-exposed-port
https://stackoverflow.com/questions/20778771/what-is-the-difference-between-0-0-0-0-127-0-0-1-and-localhost